% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccountDetail.R
\name{getAccountDetail}
\alias{getAccountDetail}
\title{Get an overview of an account on the Stellar ledger.}
\usage{
getAccountDetail(public_key)
}
\arguments{
\item{public_key}{character - your Stellar account/wallet address.}
}
\value{
list
}
\description{
Query the accounts endpoint, specifying no extra resources. Converts the JSON response to a list.
}
\note{
https://www.stellar.org/developers/horizon/reference/endpoints/accounts-single.html
}
\examples{
getAccountDetail("GCO2IP3MJNUOKS4PUDI4C7LGGMQDJGXG3COYX3WSB4HHNAHKYV5YL3VC")
}
