% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLedgerDetail.R
\name{getLedgerDetail}
\alias{getLedgerDetail}
\title{Returns a single ledger.}
\usage{
getLedgerDetail(sequence)
}
\arguments{
\item{sequence}{character - (required) a ledger id.}
}
\value{
list
}
\description{
Return a single ledger by providing a sequence id.
}
\note{
https://www.stellar.org/developers/horizon/reference/endpoints/ledgers-single.html
}
\examples{
getLedgerDetail("1")
}
