% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOperationDetail.R
\name{getOperationDetail}
\alias{getOperationDetail}
\title{Returns a single operation.}
\usage{
getOperationDetail(id)
}
\arguments{
\item{id}{character - (required) an operation id.}
}
\value{
list
}
\description{
Return a single operation by providing an id.
}
\note{
https://www.stellar.org/developers/horizon/reference/endpoints/operations-single.html
}
\examples{
\donttest{getOperationDetail("77309415424")}
}
