% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPayments.R
\name{getPayments}
\alias{getPayments}
\title{Returns all payments.}
\usage{
getPayments(cursor = NULL, limit = 10, order = "asc", data.table = TRUE,
  stream = FALSE)
}
\arguments{
\item{cursor}{numeric - optional, a paging token - where to start from. Can also be "now".}

\item{limit}{numeric - optional, the number of records to return. Default is 10.}

\item{order}{character - optional, "asc" or "desc"}

\item{data.table}{boolean - if TRUE, a data.table is returned. If FALSE or NULL, a list is returned.}

\item{stream}{boolean - if TRUE, a data.table is overwritten and the server-side streaming capability is utilised. A list will be returned.}
}
\value{
data.table or list
}
\description{
Return all payments operations that are a part of a valid transaction.
}
\note{
https://www.stellar.org/developers/horizon/reference/endpoints/payments-all.html
}
\examples{
getPayments(limit = 20)
}
