% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransactionDetail.R
\name{getTransactionDetail}
\alias{getTransactionDetail}
\title{Returns a single valid transactions.}
\usage{
getTransactionDetail(hash)
}
\arguments{
\item{hash}{character - a transaction id/hash.}
}
\value{
list
}
\description{
Return a single transaction using a transaction id.
}
\note{
https://www.stellar.org/developers/horizon/reference/endpoints/transactions-single.html
}
\examples{
\donttest{getTransactionDetail("someHash")}
}
