% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDF.R
\name{listToDF}
\alias{listToDF}
\title{Convert parsed JSON responses to a tabular format.}
\usage{
listToDF(data)
}
\arguments{
\item{data}{list}
}
\value{
data.table
}
\description{
Takes a successful response containing multiple records and converts it into a table format.
    It works by flattening semi-structured lists into a table where it can.
    For example, if there are multiple signatures to a transaction then it will warn the user before converting those values into a delimited list.
}
\examples{
txns = listToDF(stellaRbase::test_effects); summary(txns)
}
