% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_effects_acc.R
\docType{data}
\name{test_effects_acc}
\alias{test_effects_acc}
\title{Effects from a specific account}
\format{10 records}
\description{
Some sample list data returned by the getEffects_Account() function.
}
\keyword{data}
