% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_effects_desc.R
\docType{data}
\name{test_effects_desc}
\alias{test_effects_desc}
\title{Latest effects}
\format{10 records}
\description{
Some sample list data returned by the getEffects(order = "desc") function.
}
\keyword{data}
