% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_effects_txn.R
\docType{data}
\name{test_effects_txn}
\alias{test_effects_txn}
\title{Effects from a specific transaction}
\format{2 records}
\description{
Some sample list data returned by the getEffects_Transaction("samplekey") function.
}
\keyword{data}
