% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ledgers_desc.R
\docType{data}
\name{test_ledgers_desc}
\alias{test_ledgers_desc}
\title{Ledgers}
\format{10 records}
\description{
Some sample list data returned by the getLedgers(order = "desc") function.
}
\keyword{data}
