% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ledgers_stream.R
\docType{data}
\name{test_ledgers_stream}
\alias{test_ledgers_stream}
\title{Ledgers - streaming data}
\format{10 records}
\description{
Some sample list data returned by the getLedgers(stream = TRUE) function.
}
\keyword{data}
