% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_operations.R
\docType{data}
\name{test_operations}
\alias{test_operations}
\title{Operations data}
\format{10 records}
\description{
Some sample list data returned by the getOperations() function.
}
\keyword{data}
