% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_orderbook.R
\docType{data}
\name{test_orderbook_native_ca12}
\alias{test_orderbook_native_ca12}
\title{Order book response data}
\format{1 records}
\description{
A sample response from the getOrderbook() function.
}
\keyword{data}
