% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_payments_acc.R
\docType{data}
\name{test_payments_acc}
\alias{test_payments_acc}
\title{Payments data from a specific account}
\format{20 records}
\description{
Some sample list data returned by the getPayments_Account("testaccount") function.
}
\keyword{data}
