% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_txns_acc.R
\docType{data}
\name{test_txns_acc}
\alias{test_txns_acc}
\title{Transactions data from a specific account}
\format{20 records}
\description{
Some sample list data returned by the getTransactions_Account("some account") function.
}
\keyword{data}
