% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_txns_desc.R
\docType{data}
\name{test_txns_desc}
\alias{test_txns_desc}
\title{Transactions data from a specific account}
\format{10 records}
\description{
Some sample list data returned by the getTransctions(order = "desc") function.
}
\keyword{data}
